requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Browser Automation
playwright==1.41.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pypdf==3.17.4

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
speechrecognition==3.10.1

# Video Processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1
kaleido==0.2.1

# Machine Learning
scikit-learn==1.4.0
scipy==1.11.4

# Document Processing
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3